Suppress hydration warning
The date renders one way on the server, and another on the client.  This is intended because the date should be in the user's local format.  I could make the page client rendered, but want to keep the expensive mapping of alerts server-side for now.

I could put the alert mapping into an api and stop using SSR on this page

For now will keep as-is and fix later if I have time